Add function for delete from local storage
