+++ src/app.js
@@ -2,6 +2,7 @@ import './style.css';
 import {
   createlocalstorage, addBtn, listArr,
 } from './script.js';
+import { deleteTask } from './delete';
 
 const container = document.querySelector('.list-container');
 const deleteallbtn = document.querySelector('.clear');
@@ -50,15 +51,15 @@ display(listArr);
 
 const task = document.querySelectorAll('.task');
 
- export function deleteTask(idx) {
-  const storage = JSON.parse(localStorage.getItem('todos'));
-  const savedata = storage.filter((value, index) => index !== idx);
-  savedata.forEach((obj, index) => {
-    obj.index = index + 1;
-  });
-  localStorage.setItem('todos', JSON.stringify(savedata));
-  window.location.reload();
-}
+//  export function deleteTask(idx) {
+//   const storage = JSON.parse(localStorage.getItem('todos'));
+//   const savedata = storage.filter((value, index) => index !== idx);
+//   savedata.forEach((obj, index) => {
+//     obj.index = index + 1;
+//   });
+//   localStorage.setItem('todos', JSON.stringify(savedata));
+//   window.location.reload();
+// }
 
 task.forEach((element, index) => {
   const el = element.childNodes[3];
